\name{print.rvgt.range}
\alias{print.rvgt.range}
\title{Print result of range tests}
\usage{
  \method{print}{rvgt.range} (x, ...)
}
\arguments{
  \item{x}{object of class \code{"rvgt.range"} to be
  printed.}

  \item{...}{further arguments to be passed to print method
  (ignored in this function).}
}
\description{
  \code{print} method for class \code{"rvgt.range"}.
}
\details{
  Method \code{print.range.engige} tries to give a
  comprehensive overview of the test results.
}
\author{
  Josef Leydold \email{josef.leydold@wu.ac.at}
}
\seealso{
  \code{\link{rvgt.range.engine}} a description of class
  \code{"rvgt.range"}.
}

